requirements.txt
boto3==1.22.0
pandas==1.4.2
psycopg2-binary==2.9.3